# public/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AXION Industrial Operations API",
    "description": "API de missão crítica para gestão de operações industriais, chamados de manutenção e auditoria.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Servidor Base"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Ticket": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "type": { "type": "string" },
          "location": { "type": "string" },
          "priority": { "type": "string", "enum": ["Baixa", "Média", "Alta", "Crítica"] },
          "status": { "type": "string", "enum": ["Aberto", "Em atendimento", "Finalizado"] },
          "created_at": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "paths": {
    "/tickets": {
      "get": {
        "summary": "Listar tickets (com paginação e filtros)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "priority", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Lista de tickets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Ticket" } },
                    "total": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Abrir novo ticket",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["type", "location"],
                "properties": {
                  "type": { "type": "string" },
                  "location": { "type": "string" },
                  "agv_number": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Ticket criado" }
        }
      }
    },
    "/tickets/stats": {
      "get": {
        "summary": "Estatísticas em tempo real",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Objeto com contagens" }
        }
      }
    }
  }
}